message form separated into a new component, MessageField divided to Chat and MessageForm, some logic divided
