+++ JulianaJarm/src/components/MessageList/MessageList.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Message from "../Message/Message";
+
+export class MessageList extends React.Component {
+    render() {
+        return this.props.messages.map((message, index) => (
+            <Message author={ message.login } text={ message.text } key={ index } />
+        ));
+    }
+}
